+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixes & tech 🚒
 
+* App crashes on setup-selected-account screen due to incorrect icon id (#)
 * When adding new block via add-block screen, should replace current text block instead of adding a new block after this text block if this text block is empty (#325)
 * Divider block should be selectable in multi-select and scroll-and-move mode (#778)