+++ CHANGELOG.md
@@ -8,7 +8,8 @@
 
 ### Fixes & tech 🚒
 
-* App crashes on setup-selected-account screen due to incorrect icon id (#)
+* Document's image icon (uploaded from device's gallery) isn't visible in the mention suggester (#789)
+* App crashes on setup-selected-account screen due to incorrect icon id (#739)
 * When adding new block via add-block screen, should replace current text block instead of adding a new block after this text block if this text block is empty (#325)
 * Divider block should be selectable in multi-select and scroll-and-move mode (#778)