+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Fixes & tech 🚒
 
+* Remove turn-into action from page block's action menu (#787)
 * Should change number color when changing numbered block's text color (#797)
 * Document's image icon (uploaded from device's gallery) isn't visible in the mention suggester (#789)
 * App crashes on setup-selected-account screen due to incorrect icon id (#739)